+++ Projects/my_notes/lib/view/login_view.dart
@@ -32,42 +32,52 @@ class _LoginViewState extends State<LoginView> {
 
   @override
   Widget build(BuildContext context) {
-    return Column(
-      children: [
-        // email
-        TextField(
-          controller: _email,
-          autocorrect: false,
-          keyboardType: TextInputType.emailAddress,
-          decoration: CodeRepeat().hintText('xxxxx@xxxxx.xxx'),
-        ),
-        // password
-        TextField(
-          controller: _password,
-          obscureText: true,
-          autocorrect: false,
-          enableInteractiveSelection: false,
-          decoration: CodeRepeat().hintText('xxxxxxx'),
-        ),
-        TextButton(
-          onPressed: () async {
-            final email = _email.text;
-            final password = _password.text;
-            try {
-              final userCredential = await FirebaseAuth.instance
-                  .signInWithEmailAndPassword(email: email, password: password);
-              print(userCredential);
-            } on FirebaseAuthException catch (e) {
-              if (e.code == "user-not-found") {
-                print("User not found!");
-              } else if (e.code == "wrong-password") {
-                print("Wrong password!");
+    return Scaffold(
+      appBar: AppBar(title: const Text("Login")),
+      body: Column(
+        children: [
+          // email
+          TextField(
+            controller: _email,
+            autocorrect: false,
+            keyboardType: TextInputType.emailAddress,
+            decoration: CodeRepeat().hintText('xxxxx@xxxxx.xxx'),
+          ),
+          // password
+          TextField(
+            controller: _password,
+            obscureText: true,
+            autocorrect: false,
+            enableInteractiveSelection: false,
+            decoration: CodeRepeat().hintText('xxxxxxx'),
+          ),
+          TextButton(
+            onPressed: () async {
+              final email = _email.text;
+              final password = _password.text;
+              try {
+                final userCredential = await FirebaseAuth.instance
+                    .signInWithEmailAndPassword(
+                        email: email, password: password);
+                print(userCredential);
+              } on FirebaseAuthException catch (e) {
+                if (e.code == "user-not-found") {
+                  print("User not found!");
+                } else if (e.code == "wrong-password") {
+                  print("Wrong password!");
+                }
               }
-            }
-          },
-          child: const Text("Login"),
-        ),
-      ],
+            },
+            child: const Text("Login"),
+          ),
+          TextButton(
+              onPressed: () {
+                Navigator.of(context)
+                    .pushNamedAndRemoveUntil('/register/', (route) => false);
+              },
+              child: const Text('Not registered yet! Register here.'))
+        ],
+      ),
     );
   }
 }
